+++ prez/reference_data/endpoints/cql_endpoints.ttl
@@ -15,26 +15,3 @@ endpoint:get a ont:ListingEndpoint ;
     shext:offset 0 ;
 .
 
-endpoint:queryables a ont:ListingEndpoint ;
-    sh:rule [ sh:subject "?focus_node" ;
-          sh:predicate <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> ;
-          sh:object <https://prez.dev/Queryable> ] ;
-    ont:deliversClasses prez:QueryablesList ;
-    sh:target [ sh:select """SELECT DISTINCT ?focus_node
-             WHERE {
-             ?s a ?class ;
-                  ?focus_node ?o .
-              VALUES ?class {
-              dcat:Catalog
-              dcat:Dataset
-              dcat:Resource
-              skos:ConceptScheme
-              skos:Collection
-              skos:Concept
-              geo:FeatureCollection
-              geo:Feature
-              }
-                    }""" ] ;
-    shext:limit 100 ;
-    shext:offset 0 ;
-.
