+++ prez/reference_data/endpoints/endpoint_node_selection_shapes.ttl
@@ -7,6 +7,7 @@
 @prefix prof: <http://www.w3.org/ns/dx/prof/> .
 @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
 @prefix sh: <http://www.w3.org/ns/shacl#> .
+@prefix shext: <http://example.com/shacl-extension#> .
 @prefix skos: <http://www.w3.org/2004/02/skos/core#> .
 
 ex:TopLevelCatalogs
@@ -84,4 +85,29 @@ ex:Profiles
     a sh:NodeShape ;
     ont:hierarchyLevel 1 ;
     sh:targetClass prof:Profile ;
+.
+
+ex:queryables a sh:NodeShape ;
+    ont:hierarchyLevel 1 ;
+    sh:rule [ sh:subject "?focus_node" ;
+          sh:predicate <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> ;
+          sh:object <https://prez.dev/Queryable> ] ;
+    ont:deliversClasses prez:QueryablesList ;
+    sh:target [ sh:select """SELECT DISTINCT ?focus_node
+             WHERE {
+             ?s a ?class ;
+                  ?focus_node ?o .
+              VALUES ?class {
+              dcat:Catalog
+              dcat:Dataset
+              dcat:Resource
+              skos:ConceptScheme
+              skos:Collection
+              skos:Concept
+              geo:FeatureCollection
+              geo:Feature
+              }
+                    }""" ] ;
+    shext:limit 100 ;
+    shext:offset 0 ;
 .
