--- prez/reference_data/endpoints/ogc_catprez_endpoints.ttl.old
@@ -1,81 +1,0 @@
-PREFIX dcat: <http://www.w3.org/ns/dcat#>
-PREFIX dcterms: <http://purl.org/dc/terms/>
-PREFIX endpoint: <https://prez.dev/endpoint/ogccatprez/>
-PREFIX prez: <https://prez.dev/>
-PREFIX ont: <https://prez.dev/ont/>
-PREFIX sh: <http://www.w3.org/ns/shacl#>
-prefix skos: <http://www.w3.org/2004/02/skos/core#>
-PREFIX shext: <http://example.com/shacl-extension#>
-PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
-
-
-endpoint:top-level-catalog-listing a ont:ListingEndpoint ;
-    ont:endpointTemplate "/c/catalogs" ;
-    ont:deliversClasses prez:CatalogList ;  # required to determine the correct profile for ConnegP
-    sh:targetClass dcat:Catalog ;  # required for query construction
-    shext:limit 20 ;
-    shext:offset 0 ;
-    sh:target [
-      sh:select """SELECT ?focus_node
-         WHERE {
-          ?focus_node dcterms:hasPart ?child_catalog .
-          ?child_catalog a dcat:Catalog .
-            }
-         """ ] ;
-.
-
-endpoint:top-level-catalog-object a ont:ObjectEndpoint ;
-    sh:targetNode "$object" ;
-    ont:deliversClasses dcat:Catalog ;  # required for link generation for objects
-    ont:endpointTemplate "/c/catalogs/$object" ;
-    ont:parentEndpoint endpoint:top-level-catalog-listing ;
-.
-
-endpoint:lower-level-catalog-listing a ont:ListingEndpoint ;
-    ont:endpointTemplate "/c/catalogs/$parent_1/collections" ;
-    ont:deliversClasses prez:CatalogList ;  # required to determine the correct profile for ConnegP
-    ont:parentToFocusRelation dcterms:hasPart ;
-    sh:targetClass dcat:Catalog ;  # required for query construction
-    ont:parentEndpoint endpoint:top-level-catalog-object ;
-    shext:limit 20 ;
-    shext:offset 0 ;
-    sh:target [
-      sh:select """SELECT ?focus_node
-         WHERE {
-          $parent_1 dcterms:hasPart ?focus_node .
-            }
-         """ ] ;
-.
-
-endpoint:lower-level-catalog-object a ont:ObjectEndpoint ;
-    sh:targetNode "$object" ;
-    ont:deliversClasses dcat:Catalog ;  # required for link generation for objects
-    ont:endpointTemplate "/c/catalogs/$parent_1/collections/$object" ;
-    ont:parentToFocusRelation dcterms:hasPart ;
-    ont:parentEndpoint endpoint:lower-level-catalog-listing ;
-.
-
-endpoint:resource-listing a ont:ListingEndpoint ;
-    ont:endpointTemplate "/c/catalogs/$parent_2/collections/$parent_1/items" ;
-    ont:deliversClasses prez:ResourceList ;  # required to determine the correct profile for ConnegP
-    ont:parentToFocusRelation dcterms:hasPart ;
-    sh:targetClass dcat:Resource ;  # required for query construction
-    ont:parentEndpoint endpoint:lower-level-catalog-object ;
-    shext:limit 20 ;
-    shext:offset 0 ;
-    sh:target [
-      sh:select """SELECT ?focus_node
-         WHERE {
-          $parent_1 dcterms:hasPart ?focus_node .
-            }
-         """ ] ;
-.
-
-endpoint:resource-object a ont:ObjectEndpoint ;
-    sh:targetNode "$object" ;
-    ont:deliversClasses dcat:Resource ;  # required for link generation for objects
-    ont:endpointTemplate "/c/catalogs/$parent_2/collections/$parent_1/items/$object" ;
-    ont:parentToFocusRelation dcterms:hasPart ;
-    ont:parentEndpoint endpoint:resource-listing ;
-.
-
