--- prez/reference_data/endpoints/ogc_spaceprez_endpoints.ttl.old
@@ -1,78 +1,0 @@
-PREFIX dcat: <http://www.w3.org/ns/dcat#>
-PREFIX endpoint: <https://prez.dev/endpoint/spaceprez/>
-PREFIX geo: <http://www.opengis.net/ont/geosparql#>
-PREFIX ont: <https://prez.dev/ont/>
-PREFIX prez: <https://prez.dev/>
-PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
-PREFIX sh: <http://www.w3.org/ns/shacl#>
-PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
-PREFIX shext: <http://example.com/shacl-extension#>
-
-endpoint:spaceprez-home a ont:Endpoint ;
-    ont:endpointTemplate "/s" ;
-.
-
-endpoint:dataset-listing a ont:ListingEndpoint ;
-    ont:deliversClasses prez:DatasetList ;
-    sh:targetClass dcat:Dataset ;
-    ont:endpointTemplate "/s/catalogs" ;
-    shext:limit 20 ;
-    shext:offset 0 ;
-.
-
-endpoint:dataset-object a ont:ObjectEndpoint ;
-    sh:targetNode "$object" ;
-    ont:parentEndpoint endpoint:dataset-listing ;
-    ont:deliversClasses dcat:Dataset ;
-    ont:endpointTemplate "/s/catalogs/$object" ;
-.
-
-endpoint:feature-collection-listing a ont:ListingEndpoint ;
-    ont:parentEndpoint endpoint:dataset-object ;
-    sh:targetClass geo:FeatureCollection ;
-    ont:deliversClasses prez:FeatureCollectionList ;
-    ont:endpointTemplate "/s/catalogs/$parent_1/collections" ;
-    ont:parentToFocusRelation rdfs:member ;
-    shext:limit 20 ;
-    shext:offset 0 ;
-    shext:orderBy [ sh:path rdfs:label ] ;
-    sh:target [
-          sh:select """SELECT ?focus_node
-             WHERE {
-             $parent_1 rdfs:member ?focus_node .
-                }
-             """ ] ;
-.
-
-endpoint:feature-collection-object a ont:ObjectEndpoint ;
-    sh:targetNode "$object" ;
-    ont:parentEndpoint endpoint:feature-collection-listing ;
-    ont:deliversClasses geo:FeatureCollection ;
-    ont:endpointTemplate "/s/catalogs/$parent_1/collections/$object" ;
-    ont:parentToFocusRelation rdfs:member ;
-.
-
-endpoint:feature-listing a ont:ListingEndpoint ;
-    ont:parentEndpoint endpoint:feature-collection-object ;
-    sh:targetClass geo:Feature ;
-    ont:deliversClasses prez:FeatureList ;
-    ont:endpointTemplate "/s/catalogs/$parent_2/collections/$parent_1/items" ;
-    ont:parentToFocusRelation rdfs:member ;
-    shext:limit 20 ;
-    shext:offset 0 ;
-    shext:orderBy [ sh:path rdfs:label ] ;
-    sh:target [
-          sh:select """SELECT ?focus_node
-             WHERE {
-             $parent_1 rdfs:member ?focus_node .
-                }
-             """ ] ;
-.
-
-endpoint:feature-object a ont:ObjectEndpoint ;
-    sh:targetNode "$object" ;
-    ont:parentEndpoint endpoint:feature-listing ;
-    ont:deliversClasses geo:Feature ;
-    ont:endpointTemplate "/s/catalogs/$parent_2/collections/$parent_1/items/$object" ;
-    ont:parentToFocusRelation rdfs:member ;
-.
