+++ prez/repositories/__init__.py
@@ -1,0 +1,6 @@
+from .base import Repo
+from .oxrdflib import OxrdflibRepo
+from .pyoxigraph import PyoxigraphRepo
+from .remote_sparql import RemoteSparqlRepo
+
+__all__ = ["Repo", "OxrdflibRepo", "PyoxigraphRepo", "RemoteSparqlRepo"]
