+++ prez/routers/cql.py
@@ -1,8 +1,10 @@
 from typing import Optional
 
 from fastapi import APIRouter, Request, Depends
-from rdflib import URIRef
+from rdflib import Namespace
+from rdflib.namespace import URIRef
 
+from prez.reference_data.prez_ns import PREZ
 from prez.dependencies import (
     get_repo,
     cql_post_parser_dependency,
@@ -10,22 +12,24 @@ from prez.dependencies import (
     cql_get_parser_dependency,
 )
 from prez.services.listings import listing_function
-from prez.sparql.methods import Repo
+from prez.repositories import Repo
 
 router = APIRouter(tags=["ogcrecords"])
+
+OGCE = Namespace(PREZ["endpoint/extended-ogc-records/"])
 
 
 @router.post(
     path="/cql",
-    name="https://prez.dev/endpoint/cql/post",
+    name=OGCE["cql-post"],
 )
 async def cql_post_endpoint(
-    request: Request,
-    cql_parser: Optional[dict] = Depends(cql_post_parser_dependency),
-    page: int = 1,
-    per_page: int = 20,
-    repo: Repo = Depends(get_repo),
-    system_repo: Repo = Depends(get_system_repo),
+        request: Request,
+        cql_parser: Optional[dict] = Depends(cql_post_parser_dependency),
+        page: int = 1,
+        per_page: int = 20,
+        repo: Repo = Depends(get_repo),
+        system_repo: Repo = Depends(get_system_repo),
 ):
     endpoint_uri = URIRef("https://prez.dev/endpoint/cql/post")
     return await listing_function(
@@ -41,15 +45,15 @@ async def cql_post_endpoint(
 
 @router.get(
     path="/cql",
-    name="https://prez.dev/endpoint/cql/get",
+    name=OGCE["cql-get"],
 )
 async def cql_get_endpoint(
-    request: Request,
-    cql_parser: Optional[dict] = Depends(cql_get_parser_dependency),
-    page: int = 1,
-    per_page: int = 20,
-    repo: Repo = Depends(get_repo),
-    system_repo: Repo = Depends(get_system_repo),
+        request: Request,
+        cql_parser: Optional[dict] = Depends(cql_get_parser_dependency),
+        page: int = 1,
+        per_page: int = 20,
+        repo: Repo = Depends(get_repo),
+        system_repo: Repo = Depends(get_system_repo),
 ):
     endpoint_uri = URIRef("https://prez.dev/endpoint/cql/get")
     return await listing_function(
@@ -65,22 +69,23 @@ async def cql_get_endpoint(
 
 @router.get(
     path="/queryables",
-    name="https://prez.dev/endpoint/cql/queryables",
+    name=OGCE["cql-queryables"],
 )
 async def queryables_endpoint(
-    request: Request,
-    cql_parser: Optional[dict] = Depends(cql_get_parser_dependency),
-    page: int = 1,
-    per_page: int = 20,
-    repo: Repo = Depends(get_repo),
-    system_repo: Repo = Depends(get_system_repo),
+        request: Request,
+        cql_parser: Optional[dict] = Depends(cql_get_parser_dependency),
+        page: int = 1,
+        per_page: int = 20,
+        repo: Repo = Depends(get_repo),
+        system_repo: Repo = Depends(get_system_repo),
 ):
-    endpoint_uri = URIRef("https://prez.dev/endpoint/cql/queryables")
+    endpoint_uri = URIRef(request.scope.get("route").name)
     return await listing_function(
         request=request,
         repo=repo,
         system_repo=system_repo,
         endpoint_uri=endpoint_uri,
+        hierarchy_level=1,
         page=page,
         per_page=per_page,
         cql_parser=cql_parser,
