+++ prez/routers/ogc_router.py
@@ -8,7 +8,7 @@ from prez.dependencies import get_repo, get_system_repo
 from prez.services.curie_functions import get_uri_for_curie_id
 from prez.services.listings import listing_function
 from prez.services.objects import object_function
-from prez.sparql.methods import Repo
+from prez.repositories import Repo
 from prez.reference_data.prez_ns import PREZ
 from temp.grammar import IRI
