+++ prez/routers/sparql.py
@@ -11,7 +11,7 @@ from starlette.responses import StreamingResponse
 from prez.dependencies import get_repo, get_system_repo
 from prez.models.profiles_and_mediatypes import ProfilesMediatypesInfo, populate_profile_and_mediatype
 from prez.renderers.renderer import return_annotated_rdf
-from prez.sparql.methods import Repo
+from prez.repositories import Repo
 
 PREZ = Namespace("https://prez.dev/")
