+++ prez/services/listings.py
@@ -14,15 +14,15 @@ from prez.models.profiles_and_mediatypes import ProfilesMediatypesInfo, populate
 from prez.reference_data.prez_ns import PREZ
 from prez.renderers.renderer import return_from_graph
 from prez.services.link_generation import add_prez_links
-from prez.services.model_methods import get_classes
-from prez.sparql.count_query import CountQuery
-from prez.sparql.methods import Repo
-from prez.sparql.search_query import SearchQuery
+from prez.services.query_generation.classes import get_classes
+from prez.services.query_generation.count import CountQuery
+from prez.repositories import Repo
+from prez.services.query_generation.search import SearchQuery
 from temp.grammar import *
 # from rdframe.grammar import SubSelect
 # from rdframe import PrezQueryConstructor
-from temp.shacl2sparql import PrezQueryConstructor
-from temp.shacl_node_selection import NodeShape
+from prez.services.query_generation.umbrella import PrezQueryConstructor
+from prez.services.query_generation.shacl_node_selection import NodeShape
 
 log = logging.getLogger(__name__)
