+++ tests/_test_cql.py
@@ -7,7 +7,7 @@ from pyoxigraph.pyoxigraph import Store
 
 from prez.app import app
 from prez.dependencies import get_repo
-from prez.sparql.methods import Repo, PyoxigraphRepo
+from prez.repositories import Repo, PyoxigraphRepo
 from urllib.parse import quote_plus
