+++ tests/test_count.py
@@ -6,7 +6,7 @@ from pyoxigraph.pyoxigraph import Store
 
 from prez.app import app
 from prez.dependencies import get_repo
-from prez.sparql.methods import Repo, PyoxigraphRepo
+from prez.repositories import Repo, PyoxigraphRepo
 
 
 @pytest.fixture(scope="session")
