+++ tests/test_endpoints_cache.py
@@ -7,7 +7,7 @@ from rdflib import Graph
 
 from prez.app import app
 from prez.dependencies import get_repo
-from prez.sparql.methods import Repo, PyoxigraphRepo
+from prez.repositories import Repo, PyoxigraphRepo
 
 
 @pytest.fixture(scope="session")
