+++ tests/test_endpoints_management.py
@@ -8,7 +8,7 @@ from rdflib import Graph
 from prez.app import app
 from prez.dependencies import get_repo
 from prez.reference_data.prez_ns import PREZ
-from prez.sparql.methods import Repo, PyoxigraphRepo
+from prez.repositories import Repo, PyoxigraphRepo
 
 
 @pytest.fixture(scope="session")
