+++ tests/test_endpoints_vocprez.py
@@ -9,7 +9,7 @@ from rdflib.compare import isomorphic
 
 from prez.app import app
 from prez.dependencies import get_repo
-from prez.sparql.methods import Repo, PyoxigraphRepo
+from prez.repositories import Repo, PyoxigraphRepo
 
 
 @pytest.fixture(scope="session")
